+++ custom_components/danfoss_air_ccm/__init__.py
@@ -4,7 +4,12 @@ from homeassistant.core import HomeAssistant
 from .const import DOMAIN
 from .coordinator import DanfossCoordinator
 
-PLATFORMS = ["sensor", "number"]
+PLATFORMS = [
+    "sensor",
+    "number",
+    "switch",
+    "button",
+]
 
 
 async def async_setup(hass, config):
@@ -22,6 +27,8 @@ async def async_setup_entry(
     )
 
     await coordinator.async_config_entry_first_refresh()
+
+    await coordinator.async_initialize_storage()
 
     hass.data.setdefault(DOMAIN, {})[
         entry.entry_id